create entity

+++ src/main/java/com/example/storage/model/Rack.java
@@ -9,7 +9,7 @@ import java.util.List;
  * Created by Смена on 23.06.2018.
  */
 @Entity(name = "rack")
-public class Rack extends AbstractPersistable<Long> {
+public class Rack extends AbstractPersistable<Integer> {
     @Column(name = "name", length = 30)
     private String name;
